+++ src/inertial_sense_ros.cpp
@@ -1,4 +1,4 @@
-#include "inertial_sense.h"
+#include "inertial_sense_ros.h"
 #include <chrono>
 #include <stddef.h>
 #include <unistd.h>
